adds ability to input range of item IDs
This change will allow the user to input a range of item IDs.
Essentially, when a user inputs a value like "1,2-5,6", the input will
be translated into "1,2,3,4,5,6".

This can save time for the user when inputing items that are desirable
but tend to be close together in their ID (e.g. herbs and runes).